[#10392] Code Coverage
Summary:
WIP

Adding new build types: releasecov and debugcov, which are similar to release and debug but with code coverage enabled.

Make code coverage work reliably

even when proces is killed hard (SIGTERM for example, as happens
commonly for master in our java tests)

Since we are using google log, hang in there for the signal handling

Regular exits also work, thanks to clang coverage's atexit registration in
__llvm_profile_register_write_file_atexit()

Don't overwrite master/tserver runs, but create new ones with PID contained

Also gather coverage of Postgres

Now all profraw files contain information

Remove workarounds for signals, only %c

Test Plan: Jenkins: build type: debugcov

Subscribers: bogdan

Differential Revision: https://phabricator.dev.yugabyte.com/D15849

+++ managed/src/main/java/com/yugabyte/yw/common/ShellProcessHandler.java
@@ -36,6 +36,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.slf4j.Marker;
 import org.slf4j.MarkerFactory;
+import org.yb.util.BuildTypeUtil;
 
 @Singleton
 public class ShellProcessHandler {
@@ -214,7 +215,9 @@ public class ShellProcessHandler {
           LOG.error(
               "Process could not be destroyed gracefully within the specified time '{}'",
               response.description);
-          process.destroyForcibly();
+          if (!BuildTypeUtil.isCoverage()) {
+            process.destroyForcibly();
+          }
         }
       }
     } finally {
